requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
python-dotenv
openai>=1.30.0